check pay for buttons

+++ src/main/java/app/bot/controller/Chat.java
@@ -598,6 +598,8 @@ public class Chat extends TelegramLongPollingBot {
         Long chatId = update.getCallbackQuery().getFrom().getId();
         String data = update.getCallbackQuery().getData();
 
+        if(!clients.get(chatId).isPaid() && !chatId.equals(getOwnerChatId())) return;
+
         if (data.equals("5day")) {
             deleteMessage(chatId);
             executeMsg(userMessage.payUpdate(builder,chatId, 2));
